Raise an error when merging of a chunk fails
The API has been effectively ignoring potential errors raised during
execution of the bulk of the chunk merging code in `merge_chunk` -- the
part contained in a `try..except` block, specifically -- as the error
that would be contained by the corresponding `except` clause is only
logged, with the procedure competing all the way to returning. This,
combined with the caller ignoring the return value means the failure
isn't detected and the response will succeed.

This commit makes sure the error that would have been raised by the
relevant part of chunk merging code, is re-raised so that `merge_chunk`
can communicate a failure that would have to be explicitly caught or
be propagated upwards the call stack and ultimately abort the request
conventionally with Tornado flagging the response as a 500, which would
arguably be an appropriate status code for the condition indeed.