Return HTTP error when a request processing code path raises an error
The `except` block at line 1367 must deal with just about any potential
error that may be raised between it and line 1250, some of such errors
not stemming from any other catch block in between and thus being
exceptional conditions the API cannot recover from and fullfil the
request. For this reason the `except` block at line 1367 should raise
the error it catches because it doesn't really handle it to such a
degree where the request may yet be successful. The error may (unless
suppressed by another `except` block nested deeper) then finally be
caught by the `except` block at line 1376 which _always_ finishes the
response (getting `on_finish` to be called) but importantly _without_
flagging the response with an error code. Because the error it catches
may be and often is fatal for the request, this commit changes the
behaviour to have the first block raise the error after doing the
"clean-up" it does (lines 1368-1375) and the second block setting a HTTP
"server error" status code on the response (but only if another error
code hasn't already been set on it). This makes sure the otherwise
"false" positives producing successful response codes in cases where
things may have gone fatally wrong during request processing, do not
occur.