feat: add parsing of proxy headers
This commit adds the following new configuration options:
- parse_proxy_headers (bool)
- trusted_proxies (list)

parse_proxy_headers controls whether or not proxy headers will be parsed.
trusted_proxies is a list of hosts to trust when parsing the X-Forwarded-For
header, which can be either IP addresses or CIDR format ranges.

This feature leverages tornado.httpserver.HTTPServer's built-in support for
proxy headers, which are disabled by default. When enabled it makes Tornado
automatically parse values in request headers X-Real-Ip/X-Forwarded-For and
X-Scheme/X-Forwarded-Proto.
See https://www.tornadoweb.org/en/stable/httpserver.html for details.

Note that when parse_proxy_headers is enabled, clients will be able to
spoof their connecting IP address without a properly configured reverse
proxy in the middle. Additionally, trusted_proxies are evaluated by
Tornado in the parsing of X-Forwarded-For *only*, and not for X-Real-Ip
(which takes precedence) or the scheme headers.