Use `logger.exception` instead of `logger.error` in an `except` block to aid troubleshooting
The latter doesn't print the stack and thus makes troubleshooting harder.

I suppose longer-term, following Python's "easier to ask forgiveness than permission" (EAFP) "best-practice" (see PEP 463), which favours using `except` block for dealing with even more "benign" conditions such as invalid list indices, the stack trace _may_ prove to be a case of excessive verbosity in the logs -- if the error always boils down to one easily understood condition. But then -- since the `except` block is big enough it can catch a lot of things we haven't even seen before, a case can be made to _retain_ `logger.exception` so we never have to look for clues in the blind should it ever get there for _different_ reason than the aforementioned, "benign" case of some index being out of bounds. Regardless, the latter issue, when tracked down to a particular section of code where it originates, can be fixed by doing an explicit check, which would be going against the EAFP practice in favour of the "look before you leap" (LBYL) counterpart.

What I am trying to get at is that we retain `logger.exception` regardless, since the `except` block appears to cast a "wide net".