Replace `logging.error` with `loging.exception`
A benefit to this change is getting the stack trace but also including the name of the file implicated when I/O fails.

This is only superficially related to the file corruption issue I've been looking at, which I think is due to handling of the exception aborting with another exception -- but this change should at least help correlate a log record to a case of a particular file, as otherwise adjacent records cannot be related to one another with absolute certainty.