Include request identifiers with logged messages
This implements assigning of identifiers to requests, taken from the `Request-ID` HTTP header of the request if set and auto-generated otherwise, and including these with every log record where a request is implicated.

Logging of request identifiers is enabled by first adding an ID to every request handler (into the logging record formatting (not easily done the way `logging` and [possibly also] Tornado are written, unfortunately) so as to minimize changes required in the application (business logic). A base request handler class encapsulates all of the required code and only needs to be made a super-class of existing request handler classes. This is the way Tornado wants it to work, for better and for worse (I had Google Geminiassist me to arrive at the conclusion there's no other clean way to do it; Gemini wasn't used to write any code that has been committed).

Example of log messages carrying request identifiers (the number following `<module>:<line>`):

```
[E 251027 16:26:32 api:323 1] Authorization not possible: missing header
[E 251027 16:26:32 api:1322 1] Forbidden, Client not authorized for request, Authorization failed
[W 251027 16:26:32 web:2407 1] 403 GET /v1/p11/files/stream/foobar (::1) 0.64ms
[E 251027 16:26:36 api:323 2] Authorization not possible: missing header
[E 251027 16:26:36 api:1322 2] Forbidden, Client not authorized for request, Authorization failed
[W 251027 16:26:36 web:2407 2] 403 GET /v1/p11/files/stream/foobar (::1) 0.50ms
```

Unfortunately -- in my view at least -- one now has to explicitly call `super().prepare` and `super().on_finish` in any request handler subclass which has its own `prepare` or `on_finish` methods respectively -- this is OOP at its best. Better to be explicit than implicit, I guess? I tried to find another means to have requests be assigned IDs, but couldn't (with help of Google Gemini, again) find a [public]  API that wasn't `prepare` or `on_finish`. I also looked at the "execute" method in Tornado and it doesn't call anything before `prepare` that can be used, so I am fairly sure about the approach being chosen being a "lesser evil". Mind you, this has nothing to do _where_ request identifiers come from -- even if a request carries an ID of its own through an external source like a proxy forwarding the request to the API, `contextvars` is still considered the idiomatic approach for tracking the value, and it needs to be set for a request at some point -- there's no better place than `prepare`, apparently. Which necessitates `on_finish` -- to restore the value of the variable in the context.

Because `on_finish` for `RequestHandler` -- the one that restores the value of the request ID variable in the task context -- is a critical part of tracking request IDs correctly (there's a comment in the added code explaining how critical of a part it is), it MUST be called for every request, regardless, and I saw no other way to accomplish that then using `finally` (I like to argue for "idiomatic" Python, which I think it qualifies as) which in turn necessitated wrapping some bodies of code in an additional `try` block.

Unfortunately, again, due to how Git shows change hunks with `diff`, some parts of `diff` may appear confusing as the change would merely be wrapping same code in a `try` block, but `diff` shows confusing (while correct, technically) lines as it's unable to actually produce meaningful change summary.