Remove need to call `prepare` and `on_finish` [for logging request IDs] in request handlers
The odyssey to work around Tornado (and `logging`, to a smaller degree) in order to include request IDs in logged messages, continues... In an effort to absolve the developer of the relatively heavy responsibility of having to remember to call `prepare` and `on_finish` in _every_ request handler class that need request ID logging, through e.g. having `super().prepare()` and `super().on_finish()` in bodies of its own `prepare` and `on_finish`, I resorted to the `__init_subclass__` magic class method that Python happens to offer, in order to make the former "implied" or called by default. It's a fine line to walk with encapsulation of behaviour, but logging of additional contextual information should not "spill" in the form of code, into business logic layer, so I was willing to go quite far in order to maintain status-quo there. I'd go farther and have moved the entire new `tsdfileapi.api.RequestHandler` class currently supporting request ID logging, into e.g. a `tsdfile.api.logging` module, but the thing is that the class may need to be gathering other methods (`write_error` comes to mind) -- that have nothing to do with `logging` -- so it's better IMO if the class in fact stays in the `api` module, that's the best place for it somehow, also long term.

To end on a different note, transitioning to e.g. `structlog` wouldn't really help all that much everything else being equal -- we'd still need "enrichment" of context the way it has been done with the series of related commits, _or_ consider alternative where e.g. we create a logger object that encapsulates logging of request IDs without any magic of the kind I had to resort to with this series of related changes, but at the cost of needing to pass the logger around explicitly, which isn't great either to put it mildly.

At least the benefit of using `contextvars` is that we _are_ using it to its designed purpose -- _any_ logger that happens to log a message, during the lifetime of a request and more importantly _in context_ of processing of a request, will include the request identifier. Mucking about with amending Tornado's log message formatter, and using `__init_subclass__` are just details to keep business logic and the _framework_ separate so we can focus on the former and ideally _forget_ about the latter (unless "the pipes burst"). `structlog` does have some things going for it, but what I am trying to say is that I don't see how it's going to be the silver bullet alone. As long as Tornado switches the contexts in a manner which leaves us with no guarantees, for example, we're at the mercy of working either _only with_ the framework or against it. The adage "simple things should be simple and complex things should be possible" has been one I've had to reflect over working on trying to implement the feature "include request IDs in Tornado logs".