Fix tests by registering a fallback handler to serve 404
Addition of a dedicated handler class for "fallback" cases of missing endpoint registrations, done here, is to have the API serve 404 in these cases, just like before [the addition of the request ID logging feature].

With Tornado's default `tornado.web.ErrorHandler` that was effectively registered implicitly as the "fallback" handler, 404 was served for all cases of uncaptured URI patterns, but we want logging implemented with our own `RequestHandler` and yet registering that class directly produces 405 instead because Tornado is looking for [missing] `get`, `post` etc methods that correspond to the HTTP method used, and when it doesn't find them, it forces the 405 code internally. The only way to override it for _all_ methods used with a single request handler, is to raise an `tornado.web.HTTPError` from `prepare`, just as Tornado itself demonstrates with its own `tornado.web.ErrorHandler` that it delegates fallback cases to.