Add debugging utilities
This adds both utilities for aiding debugging, in the `utils` module,
but also amends the `api` module to make use of these in some choice
places that matter to me during my trying of forcing file corruption to
surface.

Since `pyresumable` is packaged separately, I should follow up by
upgrading the version of the former in which I also have made similar
amendments.