Use request itself in the processing context
Using the request itself is arguably ideal because that's where the rest
of the interesting information comes from anyway. That way only one
variable is required to obtain e.g. per-request ("requested") log level
and the assigned identifier on the request (the UUID).

This also basically extends per-request context tracking to _all_
loggers through relatively benign patching of the `Logger` class (of
Python) -- alternatives like `setLoggerClass` were experimented with but
frankly this is the best solution, ironically. The irony lies in the
fact that `logging` offers `setLoggerClass` which is an imperative
kludge to patch loggers that will be created "in the future" and which
doesn't do anything for the [already created] root logger, for much the
same reason. In short, it's not a solution to reach for.

Additionally, `isEnabledFor` makes for the best patching target because
logging levels are _cached_ inside `Logger`, so over-riding `level` with
a getter, for example, which I have also tried, creates a caching
problem because the caching mechanism is all private methods which
cannot be relied upon, so clearing the cache is discouraged through
being a hidden mechanism, but necessary because, well, `Logger`
mechanics don't actually evaluate `level` beyond caching it once, for
performance. Doing what needs to be done in `isEnabledFor` is ideal
because it side-steps all the level math, and it's a public method after
all, better yet we defer to the original `Logger.isEnabledFor` when the
request doesn't feature the header for setting custom level value.